Added methods for search for all existent entities, respecting jpa rules

+++ src/main/java/com/example/demo/service/AppointmentService.java
@@ -3,5 +3,5 @@ package com.example.demo.service;
 import com.example.demo.model.Appointment;
 
 public interface AppointmentService {
-    Appointment findById(Integer id);
+    Appointment findById(Long id);
 }
